+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Add
+- Implemened pagination for displaying a limited number of movies per page
+
 ## [0.2.0] - 2025-28-04
 ### Added
 - Login / Logout functionality